// apps/www/src/interface.ts
export interface ICountryInfo {
  id: number
  continent: string
  name: string
  alpha2: string
  alpha3: string
  slug: string
  icon: string
}

export interface ICategory {
  id: string
  name: string
  country_code: string
  category_id: string
  slug: string
  country_info: ICountryInfo
}

export interface ILeague {
  gender: string
  id: string
  country: string
  name: string
  league_id: string
  category: ICategory
  slug: string
  icon: string
  country_info?: ICountryInfo
  current_season?: ISeason
}

export interface ISeason {
  icon: string
  name: string
  id: string
  league_id: string
  start_date: string
  start_date_formatted: string
  end_date_formatted: string
  end_date: string
  season_progress: number
  year: string
  competition_id: string
  season_id: string
  slug: string
  league_name: string
  country_info: ICountryInfo
}

export interface ITeamStatistics {
  ball_possession?: number
  cards_given?: number
  corner_kicks?: number
  fouls?: number
  free_kicks?: number
  goal_kicks?: number
  injuries?: number
  offsides?: number
  red_cards?: number
  shots_blocked?: number
  shots_off_target?: number
  shots_on_target?: number
  shots_saved?: number
  shots_total?: number
  substitutions?: number
  throw_ins?: number
  yellow_cards?: number
  yellow_red_cards?: number
}

export interface ITeam {
  gender: string
  slug: string
  id: string
  team_id: string
  name: string
  short_name: string
  abbreviation: string
  icon: string
  qualifier: string
  statistics: ITeamStatistics
  players: IPlayer[]
  form: string[]
}

export interface IPlayer {
  id: string
  name: string
  rank?: number
  team: ITeam
}

export interface ISportCategorySummary {
  id: string
  name: string
  category_id: string
  competitions: ILeague[]
  competitions_count: number
  latest_competition: ILeague
  slug: string
  country_info: ICountryInfo
}

export interface ISport {
  id: string
  name: string
  sport_id: string
  slug: string
}

export interface ISportEventContext {
  sport: ISport
  season: ISeason
  competition: ILeague
  category: ICategory
}

export interface ISportEvent {
  id: string
  sport_event_id?: string
  start_time: string
  start_time_formatted: string
  start_date_formatted: string
  competitors: ITeam[]
  sport_event_context: ISportEventContext
}

export type ISportEventStatusType = 'not_started' | 'closed' | 'live'

export type ISportEventMatchStatusType =
  | 'not_started'
  | 'ended'
  | '1st_half'
  | '2nd_half'
  | 'overtime'

export interface ISportEventStatus {
  clock: {
    played: string
  }
  start_time_formatted: string
  home_score: number
  away_score: number
  status: ISportEventStatusType
  match_status: ISportEventMatchStatusType
  winner_id?: string
  match_tie?: boolean
}

export interface ISportEventStatistics {
  totals: {
    competitors: ITeam[]
  }
}

export interface ISportEventSummary {
  id: string
  sport_event_status: ISportEventStatus
  sport_event: ISportEvent
  statistics: ISportEventStatistics
}

export interface ILeagueSummary {
  league: ILeague
  sport: string
  seasons: ISeason[]
  current_season: ISeason
  summary_text: string
  current_season_teams: ITeam[]
}

export interface ISeasonSummary {
  sport_event_context: ISportEventContext
  upcoming_events: ISportEventSummary[]
  completed_events: ISportEventSummary[]
  season_summary_text?: string
  teams: ITeam[]
  best_players: IPlayer[]
  all_seasons: ISeason[]
}
// ...
export interface IVenue {
  name: string
  id: string
  venue_id: string
  capacity: number
  city_name: string
  country_name: string
  map_coordinates: string
  country_code: string
}

export interface IJersey {
  type: 'home'
}

export interface ITeamProfile {
  sport: ISport
  competitor: ITeam
  category: ICategory
  players: IPlayer[]
  manager?: IPlayer
  venue: IVenue
  summary_text?: string
  jerseys: IJersey[]
  total_players: number
  foreign_players: number
  home_players: number
}